test_video_creation.py
"""Test script for video creation functionality"""

import os
from pathlib import Path
from PIL import Image
import tempfile
from video_api import VideoGenerationAPI


def create_test_images():
    """Create simple test images"""
    images = []
    colors = [(255, 100, 100), (100, 255, 100), (100, 100, 255)]  # Red, Green, Blue
    labels = ["BEGINNING", "MIDDLE", "END"]
    
    assets_dir = Path("assets/images")
    assets_dir.mkdir(parents=True, exist_ok=True)
    
    for i, (color, label) in enumerate(zip(colors, labels)):
        # Create a simple colored image with text
        img = Image.new('RGB', (1080, 1920), color)
        
        # Save image
        img_path = assets_dir / f"test_image_{i+1}.jpg"
        img.save(img_path)
        images.append(str(img_path))
        
        print(f"Created test image: {img_path}")
    
    return images


def create_test_audio():
    """Create a simple test audio file (placeholder)"""
    audio_dir = Path("assets/audio")
    audio_dir.mkdir(parents=True, exist_ok=True)
    
    # Create an empty MP3 file as placeholder
    audio_path = audio_dir / "test_audio.mp3"
    
    # Note: This creates an empty file. For real testing, you'd need actual audio
    with open(audio_path, 'wb') as f:
        f.write(b'')  # Empty file
    
    print(f"Created placeholder audio file: {audio_path}")
    return str(audio_path)


def test_video_api():
    """Test the video generation API with sample data"""
    
    print("Testing Video Generation API")
    print("=" * 40)
    
    # Create test assets
    test_images = create_test_images()
    test_audio = create_test_audio()
    
    # Sample timing data
    timing_data = [
        {"start": 0.0, "end": 2.0, "text": "This is the beginning of our story"},
        {"start": 2.0, "end": 4.0, "text": "Now we're in the middle section"},
        {"start": 4.0, "end": 6.0, "text": "And here's how it all ends"}
    ]
    
    # Initialize API
    api = VideoGenerationAPI()
    
    # Test input validation
    print("\n1. Testing input validation...")
    
    try:
        # This should fail - wrong number of images
        result = api.create_video(
            images=test_images[:2],  # Only 2 images
            audio=test_audio,
            timing_data=timing_data
        )
        print("ERROR: Should have failed with wrong image count")
    except ValueError as e:
        print(f"✓ Correctly caught error: {e}")
    
    # Test sample timing data generation
    print("\n2. Testing sample timing data generation...")
    
    sample_texts = [
        "First segment of text",
        "Second segment of text", 
        "Final segment of text"
    ]
    
    generated_timing = api.create_sample_timing_data(sample_texts, 6.0)
    print(f"✓ Generated timing data: {generated_timing}")
    
    # Test video creation (will fail due to empty audio file, but tests the pipeline)
    print("\n3. Testing video creation...")
    
    try:
        result = api.create_video(
            images=test_images,
            audio=test_audio,
            timing_data=timing_data,
            duration=6.0
        )
        
        if result["status"] == "success":
            print(f"✓ Video created successfully: {result['video_path']}")
        else:
            print(f"⚠ Video creation failed (expected due to empty audio): {result['error']}")
            
    except Exception as e:
        print(f"⚠ Video creation failed (expected due to missing dependencies): {e}")
    
    print("\n4. Summary")
    print("- Created test images ✓")
    print("- Created placeholder audio ✓") 
    print("- API validation working ✓")
    print("- Timing data generation working ✓")
    print("- Video creation pipeline ready (needs real audio + FFmpeg)")


if __name__ == "__main__":
    test_video_creation()